feat: !roles without arguments now returns a list of assignable roles; !settings all lists all available settings; original message and the bot response are now deleted simultaneously when assigning or listing roles

+++ src/modules/roles.js
@@ -71,6 +71,7 @@ module.exports = function(bot) {
 	bot.on('messageCreate', msg => {
 		if (! msg.member) return;
 		if (msg.author.bot) return;
+		if (msg.content === '!roles') return;
 		if (msg.content[0] === '!') return deleteOtherMessage(msg);
 
 		const guild = msg.channel.guild;
@@ -103,94 +104,126 @@ module.exports = function(bot) {
 						const removedStr = (removedRoleNames.length === 1 ? 'role' : 'roles');
 
 						let msgParts = [];
-						if (addedRoleNames.length > 0) msgParts.push(`added ${addedStr} ${util.prettyList(addedRoleNames)}`);
-						if (removedRoleNames.length > 0) msgParts.push(`removed ${removedStr} ${util.prettyList(removedRoleNames)}`);
+						if (addedRoleNames.length > 0) msgParts.push(`now have the ${addedStr} ${util.prettyList(addedRoleNames)}`);
+						if (removedRoleNames.length > 0) msgParts.push(`no longer have the ${removedStr} ${util.prettyList(removedRoleNames)}`);
 
-						return bot.createMessage(msg.channel.id, `${msg.author.mention} ${msgParts.join(' and ')}`);
+						return bot.createMessage(msg.channel.id, `${msg.author.mention} You ${msgParts.join(' and ')}`);
 					}, (e) => {console.error(e);})
 					.then(doneMsg => {
-						settings.getMultiple(guild.id, ['roles.autoDeleteMessages', 'roles.autoDeleteOtherMessages', 'roles.autoDeleteResponseDelay']).then(values => {
+						settings.getMultiple(guild.id, ['roles.autoDeleteMessages', 'roles.autoDeleteOtherMessages', 'roles.autoDeleteDelay']).then(values => {
 							if (! values['roles.autoDeleteMessages']) return;
 
-							bot.deleteMessage(msg.channel.id, msg.id);
-
-							const selfDeleteDelay = parseInt(values['roles.autoDeleteResponseDelay'], 10) * 1000;
-							setTimeout(() => bot.deleteMessage(msg.channel.id, doneMsg.id), selfDeleteDelay);
+							const deleteDelay = parseInt(values['roles.autoDeleteDelay'], 10) * 1000;
+							setTimeout(() => {
+								bot.deleteMessage(msg.channel.id, msg.id);
+								bot.deleteMessage(msg.channel.id, doneMsg.id);
+							}, deleteDelay);
 						});
 					});
 				});
 		});
 	});
 
-	// (ADMIN) Setting assignable roles
+	// Assignable roles
 	bot.registerCommand('roles', (msg, args) => {
 		if (! msg.member) return;
-		if (! msg.member.permission.has('administrator')) return;
 
 		const guild = msg.channel.guild;
 
-		const roleMods = getModsFromText(args.join(' '));
-		let toAdd = findRolesByName(guild, roleMods.add).map(role => role.id);
-		let toRemove = findRolesByName(guild, roleMods.remove).map(role => role.id);
+		if (args.length === 0) {
+			getChannels(guild).then(channels => {
+				if (channels.indexOf(msg.channel.id) === -1) return;
 
-		if (toAdd.length === 0 && toRemove.length === 0) return;
-
-		AssignableRole.query()
-			.where('guild_id', guild.id)
-			.then(assignableRoles => {
-				const assignableRolesByRoleId = util.indexBy(assignableRoles, 'role_id');
-
-				// Remove duplicate roles
-				toAdd = toAdd.filter(roleId => ! assignableRolesByRoleId[roleId]);
-
-				// Cleanup: add non-existing assignable roles to the toRemove list
-				toRemove = assignableRoles.reduce((toRemove, assignableRole) => {
-					if (guild.roles.get(assignableRole.role_id) === null) {
-						toRemove.push(assignableRole.role_id);
-					}
-
-					return toRemove;
-				}, toRemove);
-
-				toAdd = util.unique(toAdd);
-				toRemove = util.unique(toRemove);
-
-				const promises = [];
-
-				if (toAdd.length > 0) {
-					toAdd.forEach(roleId => {
-						let q = AssignableRole.query()
-							.insert({
-								guild_id: guild.id,
-								role_id: roleId
-							});
-
-						promises.push(q);
-					});
-				}
-
-				if (toRemove.length > 0) {
-					let q = AssignableRole.query()
-						.where('guild_id', guild.id)
-						.whereIn('role_id', toRemove)
-						.delete();
-
-					promises.push(q);
-				}
-
-				Promise.all(promises)
-					.then(() => {
-						return AssignableRole.query().where('guild_id', guild.id);
-					})
-					.then(result => {
-						const assignableRoleNames = result.map(assignableRole => {
+				// Get assignable roles
+				AssignableRole.query()
+					.where('guild_id', guild.id)
+					.then(assignableRoles => {
+						const assignableRoleNames = assignableRoles.map(assignableRole => {
 							const role = guild.roles.get(assignableRole.role_id);
-							return (role ? role.name : `[UNKNOWN ROLE ${assignableRole.role_id}]`);
+							return (role ? `**${role.name}**` : `[UNKNOWN ROLE ${assignableRole.role_id}]`);
 						});
 
-						bot.createMessage(msg.channel.id, `Roles updated! Assignable roles are now: ${util.prettyList(assignableRoleNames)}`);
+						bot.createMessage(msg.channel.id, `The following roles are available: ${util.prettyList(assignableRoleNames)}`)
+							.then(listMsg => {
+								settings.getMultiple(guild.id, ['roles.autoDeleteMessages', 'roles.autoDeleteListDelay']).then(values => {
+									if (! values['roles.autoDeleteMessages']) return;
+
+									const deleteDelay = parseInt(values['roles.autoDeleteListDelay'], 10) * 1000;
+									setTimeout(() => {
+										bot.deleteMessage(msg.channel.id, msg.id);
+										bot.deleteMessage(msg.channel.id, listMsg.id);
+									}, deleteDelay);
+								});
+							});
 					});
 			});
+		} else {
+			// (ADMIN) Set assignable roles
+			if (! msg.member.permission.has('administrator')) return;
+
+			const roleMods = getModsFromText(args.join(' '));
+			let toAdd = findRolesByName(guild, roleMods.add).map(role => role.id);
+			let toRemove = findRolesByName(guild, roleMods.remove).map(role => role.id);
+
+			if (toAdd.length === 0 && toRemove.length === 0) return;
+
+			AssignableRole.query()
+				.where('guild_id', guild.id)
+				.then(assignableRoles => {
+					const assignableRolesByRoleId = util.indexBy(assignableRoles, 'role_id');
+
+					// Remove duplicate roles
+					toAdd = toAdd.filter(roleId => ! assignableRolesByRoleId[roleId]);
+
+					// Cleanup: add non-existing assignable roles to the toRemove list
+					toRemove = assignableRoles.reduce((toRemove, assignableRole) => {
+						if (guild.roles.get(assignableRole.role_id) === null) {
+							toRemove.push(assignableRole.role_id);
+						}
+
+						return toRemove;
+					}, toRemove);
+
+					toAdd = util.unique(toAdd);
+					toRemove = util.unique(toRemove);
+
+					const promises = [];
+
+					if (toAdd.length > 0) {
+						toAdd.forEach(roleId => {
+							let q = AssignableRole.query()
+								.insert({
+									guild_id: guild.id,
+									role_id: roleId
+								});
+
+							promises.push(q);
+						});
+					}
+
+					if (toRemove.length > 0) {
+						let q = AssignableRole.query()
+							.where('guild_id', guild.id)
+							.whereIn('role_id', toRemove)
+							.delete();
+
+						promises.push(q);
+					}
+
+					Promise.all(promises)
+						.then(() => {
+							return AssignableRole.query().where('guild_id', guild.id);
+						})
+						.then(result => {
+							const assignableRoleNames = result.map(assignableRole => {
+								const role = guild.roles.get(assignableRole.role_id);
+								return (role ? role.name : `[UNKNOWN ROLE ${assignableRole.role_id}]`);
+							});
+
+							bot.createMessage(msg.channel.id, `Roles updated! Assignable roles are now: ${util.prettyList(assignableRoleNames)}`);
+						});
+				});
+		}
 	});
 
 	// (ADMIN) Set channels where roles can be assigned
